refactor(agent): assemble 并入 prompts，统一 prompt 组装域
assembleSystemContext（动态 system 上下文装配）与 PROMPT_TEMPLATES（静态 user-turn 模板）同属
「构造发给模型的 prompt」，合并到 prompts.ts 一处：

- assemble.ts 的类型（AssemblePrMeta / AssembleSessionSnapshot / AssembleInput）与
  assembleSystemContext + 各 render 辅助函数并入 prompts.ts，删除 assemble.ts。
- orchestrator / planner / index 改从 prompts 引用；公共 API（经 index 再导出）不变。

纯结构调整、行为不变；agent 49 tests / lint / typecheck / desktop build 全绿。
